requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.46.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
black==25.1.0
boto3==1.36.26
botocore==1.36.26
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
Crawl4AI==0.4.248
cryptography==44.0.1
cssselect==1.2.0
distro==1.9.0
docstring_parser==0.16
fake-http-header==0.3.5
fake-useragent==2.0.3
filelock==3.17.0
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.6.1
isort==6.0.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
litellm==1.61.13
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
nltk==3.9.1
numpy==2.2.3
openai==1.63.2
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
peewee==3.17.9
phidata==2.7.10
pillow==10.4.0
platformdirs==4.3.6
playwright==1.50.0
propcache==0.3.0
psutil==7.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pyee==12.1.1
Pygments==2.19.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
s3transfer==0.11.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
SQLAlchemy==2.0.38
tavily-python==0.5.1
tf-playwright-stealth==1.1.1
tiktoken==0.9.0
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
yfinance==0.2.54
zipp==3.21.0
